split into components, update og image
